requirements.txt
torch
torchvision
transformers
clip-by-openai
opencv-python
scikit-learn
numpy
scipy
matplotlib
seaborn
tqdm
pillow
moviepy
faiss-cpu
huggingface-hub
datasets
accelerate
wandb
pyyaml
google-generativeai 